data-importer-service/requirements.txt
pandas
mysql-connector-python 